+++ tests/Replication2/ReplicatedState/StateCleanupTest.cpp
@@ -25,7 +25,7 @@
 #include "Replication2/ReplicatedLog/TestHelper.h"
 
 #include "Replication2/ReplicatedState/ReplicatedState.h"
-#include "Replication2/ReplicatedState/ReplicatedState.tpp"
+#include "Replication2/ReplicatedState/ReplicatedStateImpl.tpp"
 #include "Replication2/ReplicatedState/ReplicatedStateMetrics.h"
 #include "Replication2/Streams/LogMultiplexer.h"
 #include "Replication2/Mocks/FakeReplicatedState.h"
