+++ tests/Replication2/ReplicatedState/StateMachines/DocumentState/DocumentStateMachineTest.h
@@ -28,7 +28,7 @@
 // errors.
 #include "utils/string.hpp"
 
-#include "Replication2/ReplicatedState/ReplicatedState.tpp"
+#include "Replication2/ReplicatedState/ReplicatedStateImpl.tpp"
 #include "Replication2/ReplicatedState/ReplicatedStateFeature.h"
 
 #include "Mocks/Servers.h"
